requirements.txt
PyMuPDF
pytesseract
spacy
python-docx
ultralytics
Pillow
opencv-python
fastapi
uvicorn
python-multipart